Revising codebase to make operations more clear.
That means revising most of the 3-dimensional usages of "y" back into "z". That also means clarifying "length" & "width" to indicate whether it is length on the "x" or "z" dimension. Then I have to unscramble all the jury-rigged fixes I made in order to work around the problems this terrible system made.