Fixed ongoing missing libraries issue
The project, when opened in JMonkey, had been complaining about missing
libraries for some time. Mostly due to my inexperience, I had just
avoided resolving them.

But, now that I'm revisiting this project and I actually have some
modicum of experience, I've corrected the issue.

I've removed the unecessary/outdated requirements and added a new lib
directory. The jme3 libs were automatically relocated to this directory.

In addition, a new and updated version of the JSON library has been
added. Instead of keeping it in an undisclosed location (that I totally
never lost track of), it has been added to the repository (complete with
license).